requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
pydantic-settings>=2.1,<3
SQLAlchemy>=2.0,<3
psycopg[binary]>=3.1,<4
google-genai>=1.0,<2
httpx>=0.27,<1
